+++ config.js
@@ -21,3 +21,16 @@ exports.node_options = function options() {
 exports.files = function files() {
   return getConfiguration().files;
 };
+
+exports.subdirectory = function subdirectory() {
+  return getConfiguration().subdirectory;
+};
+
+exports.requires = function requires() {
+  const files = getConfiguration().requires || [];
+
+  if(!Array.isArray(files))
+    throw new Error("mocha.requires configuration must be an array of files");
+
+  return files.map(s => s.toString());
+};