+++ worker/runtest.js
@@ -12,6 +12,9 @@ const
   args = JSON.parse(process.argv[process.argv.length - 1]),
   options = args.options;
 
+for(let file of args.requires)
+  require(file);
+
 if (Object.keys(options || {}).length) {
   console.log(`Applying Mocha options:\n${indent(JSON.stringify(options, null, 2))}`);
 } else {