+++ CHANGELOG.md
@@ -24,7 +24,7 @@ Pays attention to PATHEXT environment variable.
 - Feature: Rerun failed tests
 - Feature: Rerun last set of tests
 - Feature: Dump severe error to output channel
-- Fix: When selecting tests, it did not use Mocha options in workspace settings
+- Fix: When selecting tests, it did not use Mocha options in  settings
 
 0.0.1 (2016-04-25)
 =====